Implementation of routes authorization

+++ src/categories/categories.controller.ts
@@ -9,9 +9,15 @@ import {
   Delete,
   UsePipes,
   ValidationPipe,
+  UseGuards,
+  UnauthorizedException,
+  ParseIntPipe,
 } from '@nestjs/common';
 import { CategoryService } from './categories.service';
 import { CreateCategoryDto } from './dto/createCategoryDto';
+import { JwtAuthGuard } from 'src/posts/posts.controller';
+import { CurrentUser } from 'src/auth/decorator/current-user.decorator';
+import { JwtPayload } from 'src/auth/interface/jwt-payload.interface';
 
 @Controller('categories')
 export class CategoryController {
@@ -28,19 +34,44 @@ export class CategoryController {
   }
 
   @Post()
+  @UseGuards(JwtAuthGuard)
   @UsePipes(new ValidationPipe())
-  create(@Body() categoryDto: CreateCategoryDto) {
+  async create(
+    @Body() categoryDto: CreateCategoryDto,
+    @CurrentUser() user: JwtPayload,
+  ) {
+    // Check if the user is an admin before allowing category creation
+    if (!user.isAdmin) {
+      throw new UnauthorizedException('Only admins can create categories');
+    }
     return this.categoryService.create(categoryDto);
   }
 
   @Put(':id')
+  @UseGuards(JwtAuthGuard)
   @UsePipes(new ValidationPipe())
-  update(@Param('id') id: string, @Body() categoryDto: CreateCategoryDto) {
+  update(
+    @Param('id', ParseIntPipe) id: string,
+    @Body() categoryDto: CreateCategoryDto,
+    @CurrentUser() user: JwtPayload,
+  ) {
+    // Check if the user is an admin before allowing category update
+    if (!user.isAdmin) {
+      throw new UnauthorizedException('Only admins can update categories');
+    }
     return this.categoryService.update(+id, categoryDto);
   }
 
   @Delete(':id')
-  remove(@Param('id') id: string) {
+  @UseGuards(JwtAuthGuard)
+  remove(
+    @Param('id', ParseIntPipe) id: string,
+    @CurrentUser() user: JwtPayload,
+  ) {
+    // Check if the user is an admin before allowing category removal
+    if (!user.isAdmin) {
+      throw new UnauthorizedException('Only admins can delete categories');
+    }
     return this.categoryService.remove(+id);
   }
 }
