+++ src/categories/categories.service.ts
@@ -1,5 +1,5 @@
 // src/services/category.service.ts
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { Category } from '../entities/category.entity';
@@ -29,11 +29,28 @@ export class CategoryService {
   }
 
   async update(id: number, categoryDto: CreateCategoryDto): Promise<Category> {
-    await this.categoryRepository.update(id, categoryDto);
+    const existingCategory = await this.categoryRepository.findOne({
+      where: { id: id },
+    });
+    if (!existingCategory) {
+      throw new HttpException('Category not found', HttpStatus.BAD_REQUEST);
+    }
+    // Update category properties
+    existingCategory.name = categoryDto.name || existingCategory.name;
+
+    // Save the updated post
+    await this.categoryRepository.save(existingCategory);
+
     return this.findOne(id);
   }
 
   async remove(id: number): Promise<void> {
+    const existingCategory = await this.categoryRepository.findOne({
+      where: { id: id },
+    });
+    if (!existingCategory) {
+      throw new HttpException('Category not found', HttpStatus.BAD_REQUEST);
+    }
     await this.categoryRepository.delete(id);
   }
 }
